+++ lib/pgkeeper/web.rb
@@ -13,6 +13,7 @@ require "pgkeeper/web/jobs"
 require "pgkeeper/web/actions"
 require "pgkeeper/web/dashboard"
 require "pgkeeper/web/connections"
+require "pgkeeper/web/database_admin"
 require "pgkeeper/web/app"
 
 module PgKeeper